fixed typo

+++ data/things.py
@@ -31,7 +31,7 @@ class THINGSTriplet(torch.utils.data.Dataset):
     
         
         with open(os.path.join(self.root, 'triplets', 'train_90.npy' if train else 'test_10.npy'), 'rb') as f:
-            self.triplets = np.load(f)
+            self.triplets = np.load(f).astype(int)
         
         if download:
             f = urllib.request.urlopen(object_concepts_link)
